% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\docType{class}
\name{LogisticRegression}
\alias{LogisticRegression}
\title{LogisticRegression Class}
\description{
Implements a multinomial logistic regression model using R6.
}
\details{
#' @section Fields:
\describe{
  \item{\code{data}}{Data frame containing the raw data loaded from the specified file.}
  \item{\code{missing_values}}{List or vector indicating missing values in each column.}
  \item{\code{missing_values_percent}}{Percentage of missing values per column.}
  \item{\code{cols_names}}{Character vector containing the names of all columns in the dataset.}
  \item{\code{cat_cols_names}}{Character vector containing the names of categorical columns.}
  \item{\code{target}}{Character indicating the name of the target variable for regression.}
  \item{\code{predictors}}{Character vector containing the names of predictor variables used for the model.}
  \item{\code{prepared_data}}{List containing the processed data after preparation (encoding, normalization, etc.).}
  \item{\code{X}}{Matrix of predictor variables used for model training.}
  \item{\code{y}}{Vector of target variable labels corresponding to the observations.}
  \item{\code{n}}{Total number of observations in the dataset.}
  \item{\code{X_train}}{Matrix of predictor variables for the training set.}
  \item{\code{y_train}}{Vector of target variable labels for the training set.}
  \item{\code{X_test}}{Matrix of predictor variables for the test set.}
  \item{\code{y_test}}{Vector of target variable labels for the test set.}
  \item{\code{predicted_targets}}{Data frame containing the classes predicted by the model on the test set.}
  \item{\code{accuracy}}{Numeric value indicating the accuracy of the model on the test set.}
  \item{\code{coefficients}}{Vector or matrix containing the coefficients of the logistic regression model.}
  \item{\code{levels_map}}{List or vector mapping the levels of the encoded categorical variables.}
  \item{\code{class_labels}}{Character vector containing the labels of the different classes.}
  \item{\code{class_frequencies}}{Vector or table indicating the frequency of each class in the training data.}
}
}
\section{Methods}{

\describe{
  \item{\code{initialize(file_path, delimiter)}}{Initializes the model with data from a CSV or Excel file.}
  \item{\code{handle_missing_values(num_method, cat_method)}}{Handles missing values in the dataset.}
  \item{\code{auto_select_target(entropy_threshold, correlation_threshold, weight_entropy, weight_correlation)}}{Automatically selects the target variable based on entropy and correlation thresholds.}
  \item{\code{auto_remove_columns(correlation_threshold)}}{Automatically removes columns with high correlation.}
  \item{\code{prepare_data(target, columns_to_remove, test_size)}}{Prepares the data for training by defining predictors and splitting into training and testing sets.}
  \item{\code{fit(learning_rate, max_iter, batch_size, tol)}}{Fits the logistic regression model using gradient descent.}
  \item{\code{predict()}}{Predicts classes for the test set and returns accuracy.}
  \item{\code{predict_proba()}}{Predicts class probabilities for the test set.}
  \item{\code{summary()}}{Prints a summary of the model.}
  \item{\code{print()}}{Prints basic information about the model.}
}
}

\examples{

## ------------------------------------------------
## Method `LogisticRegression$new`
## ------------------------------------------------

\dontrun{
# Initialize the model with a CSV file and comma delimiter
model <- LogisticRegression$new("data/loan_data.csv", ",")
}
Handle Missing Values in the Dataset

Processes missing values in the dataset using specified methods for numerical and categorical variables.


## ------------------------------------------------
## Method `LogisticRegression$handle_missing_values`
## ------------------------------------------------

\dontrun{
# Handle missing values by replacing numerical NAs with mean and categorical NAs with mode
model$handle_missing_values(num_method = "mean", cat_method = "mode")
}
Automatically Select Target Variable

Automatically selects the target variable based on entropy and correlation thresholds.


## ------------------------------------------------
## Method `LogisticRegression$auto_select_target`
## ------------------------------------------------

\dontrun{
# Automatically select target with specified thresholds
model$auto_select_target(entropy_threshold = 0.5, correlation_threshold = 0.3, 
                        weight_entropy = 0.7, weight_correlation = 0.3)
}
Automatically remove unnecessary columns

Automatically removes columns with high correlation based on the specified threshold.


## ------------------------------------------------
## Method `LogisticRegression$auto_remove_columns`
## ------------------------------------------------

\dontrun{
# Remove columns with average correlation above 0.9
columns_removed <- model$auto_remove_columns(correlation_threshold = 0.9)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.


## ------------------------------------------------
## Method `LogisticRegression$prepare_data`
## ------------------------------------------------

\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Fit the Logistic Regression Model

Adjusts the model coefficients using gradient descent.


## ------------------------------------------------
## Method `LogisticRegression$fit`
## ------------------------------------------------

\dontrun{
model <- LogisticRegression$new("data/loan_data.csv", ",")
model$prepare_data(target = "default", columns_to_remove = NULL, test_size = 0.2)
model$fit(learning_rate = 0.01, max_iter = 1000, batch_size = 32, tol = 1e-4)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.


## ------------------------------------------------
## Method `LogisticRegression$predict`
## ------------------------------------------------

\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.


## ------------------------------------------------
## Method `LogisticRegression$predict_proba`
## ------------------------------------------------

\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Summary of the Logistic Regression Model

Prints a summary of the logistic regression model, including the number of observations, predictors, classes, class labels, class frequencies, and model coefficients.


## ------------------------------------------------
## Method `LogisticRegression$summary`
## ------------------------------------------------

\dontrun{
# Print the model summary
model$summary()
}
Print basics informations about the Logistic Regression Model

Prints basic information about the logistic regression model, including the number of observations, predictors, classes, class labels, class frequencies, and model coefficients.


## ------------------------------------------------
## Method `LogisticRegression$print`
## ------------------------------------------------

\dontrun{
# Print the model information
model$print()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogisticRegression-new}{\code{LogisticRegression$new()}}
\item \href{#method-LogisticRegression-handle_missing_values}{\code{LogisticRegression$handle_missing_values()}}
\item \href{#method-LogisticRegression-auto_select_target}{\code{LogisticRegression$auto_select_target()}}
\item \href{#method-LogisticRegression-auto_remove_columns}{\code{LogisticRegression$auto_remove_columns()}}
\item \href{#method-LogisticRegression-prepare_data}{\code{LogisticRegression$prepare_data()}}
\item \href{#method-LogisticRegression-fit}{\code{LogisticRegression$fit()}}
\item \href{#method-LogisticRegression-predict}{\code{LogisticRegression$predict()}}
\item \href{#method-LogisticRegression-predict_proba}{\code{LogisticRegression$predict_proba()}}
\item \href{#method-LogisticRegression-summary}{\code{LogisticRegression$summary()}}
\item \href{#method-LogisticRegression-print}{\code{LogisticRegression$print()}}
\item \href{#method-LogisticRegression-clone}{\code{LogisticRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$new(file_path, delimiter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Character. The path to the data file (CSV or Excel).}

\item{\code{delimiter}}{Character. The delimiter used in the CSV file (e.g., "," for comma).}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Initialize the model with a CSV file and comma delimiter
model <- LogisticRegression$new("data/loan_data.csv", ",")
}
Handle Missing Values in the Dataset

Processes missing values in the dataset using specified methods for numerical and categorical variables.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-handle_missing_values"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-handle_missing_values}{}}}
\subsection{Method \code{handle_missing_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$handle_missing_values(
  num_method = c("none", "mean", "median", "mode", "remove"),
  cat_method = c("none", "mode", "remove")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_method}}{Character. Method to handle missing numerical values. Options are:
\itemize{
  \item \code{"none"}: Do not handle missing values.
  \item \code{"mean"}: Replace missing values with the mean of the column.
  \item \code{"median"}: Replace missing values with the median of the column.
  \item \code{"mode"}: Replace missing values with the mode of the column.
  \item \code{"remove"}: Remove rows with missing numerical values.
}}

\item{\code{cat_method}}{Character. Method to handle missing categorical values. Options are:
\itemize{
  \item \code{"none"}: Do not handle missing values.
  \item \code{"mode"}: Replace missing values with the mode of the column.
  \item \code{"remove"}: Remove rows with missing categorical values.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Handle missing values by replacing numerical NAs with mean and categorical NAs with mode
model$handle_missing_values(num_method = "mean", cat_method = "mode")
}
Automatically Select Target Variable

Automatically selects the target variable based on entropy and correlation thresholds.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-auto_select_target"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-auto_select_target}{}}}
\subsection{Method \code{auto_select_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$auto_select_target(
  entropy_threshold = 0.5,
  correlation_threshold = 0.5,
  weight_entropy = 0.5,
  weight_correlation = 0.5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entropy_threshold}}{Numeric. Threshold for entropy to consider a variable as a potential target.}

\item{\code{correlation_threshold}}{Numeric. Threshold for correlation to filter out highly correlated variables.}

\item{\code{weight_entropy}}{Numeric. Weight assigned to entropy in the selection criteria.}

\item{\code{weight_correlation}}{Numeric. Weight assigned to correlation in the selection criteria.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the name of the selected target variable.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Automatically select target with specified thresholds
model$auto_select_target(entropy_threshold = 0.5, correlation_threshold = 0.3, 
                        weight_entropy = 0.7, weight_correlation = 0.3)
}
Automatically remove unnecessary columns

Automatically removes columns with high correlation based on the specified threshold.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-auto_remove_columns"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-auto_remove_columns}{}}}
\subsection{Method \code{auto_remove_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$auto_remove_columns(correlation_threshold = 0.9)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{correlation_threshold}}{Numeric. Threshold for average Cramér's V correlation to consider a column for removal (default is 0.9).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of column names to remove or \code{NULL} if no columns meet the criteria.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Remove columns with average correlation above 0.9
columns_removed <- model$auto_remove_columns(correlation_threshold = 0.9)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$prepare_data(target, columns_to_remove, test_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Character. The name of the target variable.}

\item{\code{columns_to_remove}}{Character vector. Names of columns to remove from the predictors.}

\item{\code{test_size}}{Numeric. Proportion of the dataset to include in the test split (e.g., 0.3 for 30%).}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Fit the Logistic Regression Model

Adjusts the model coefficients using gradient descent.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-fit"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$fit(learning_rate, max_iter, batch_size, tol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{Numeric. The learning rate for gradient descent.}

\item{\code{max_iter}}{Integer. The maximum number of iterations.}

\item{\code{batch_size}}{Integer. The size of each mini-batch.}

\item{\code{tol}}{Numeric. The tolerance for convergence.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the fitted coefficients.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
model <- LogisticRegression$new("data/loan_data.csv", ",")
model$prepare_data(target = "default", columns_to_remove = NULL, test_size = 0.2)
model$fit(learning_rate = 0.01, max_iter = 1000, batch_size = 32, tol = 1e-4)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-predict"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Character. The name of the target variable.}

\item{\code{columns_to_remove}}{Character vector. Names of columns to remove from the predictors.}

\item{\code{test_size}}{Numeric. Proportion of the dataset to include in the test split (e.g., 0.2 for 20%).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a list containing the prepared data, predictors, and training/testing splits.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Prepare Data for Modeling

Prepares the dataset by defining the target variable, removing specified columns, encoding categorical variables, normalizing numerical variables, and splitting the data into training and testing sets.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict_proba()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Character. The name of the target variable.}

\item{\code{columns_to_remove}}{Character vector. Names of columns to remove from the predictors.}

\item{\code{test_size}}{Numeric. Proportion of the dataset to include in the test split (e.g., 0.2 for 20%).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a list containing the prepared data, predictors, and training/testing splits.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Prepare the data with a specified target, columns to remove, and test size
model$prepare_data(target = "default", columns_to_remove = c("id", "timestamp"), test_size = 0.3)
}
Summary of the Logistic Regression Model

Prints a summary of the logistic regression model, including the number of observations, predictors, classes, class labels, class frequencies, and model coefficients.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-summary"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None. Prints the summary to the console.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Print the model summary
model$summary()
}
Print basics informations about the Logistic Regression Model

Prints basic information about the logistic regression model, including the number of observations, predictors, classes, class labels, class frequencies, and model coefficients.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-print"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None. The method prints information to the console.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Print the model information
model$print()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
