% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{summary.R6}
\alias{summary.R6}
\title{S3 Method for summary of R6 Objects}
\usage{
\method{summary}{R6}(object, ...)
}
\arguments{
\item{object}{An R6 object to summarize.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
Provides a summary method for objects of R6 classes.
}
